DHT22 WORKING!!!
currently uses special registration in Index, and only dht works, but that is enough.

Disclaimer: Every sent pin change on delaying pins must have a response.  This can be a dummy response (AvrCPU.setpinstate to current), or a custom Timing packet as needed, but it needs a response, or the entire thing will hang.