added raf to sg90, still not helping tho
added better pwm to rgb. still not perfect. think i need to determine period and then not touch it afterwords, because there seems to be an issue where if youre not constantly pulsing it, then the period will get messed up. so add check and dont constantly recompute the period