This modification is being implemented as a result of considerations that, while not explicitly defined, are nonetheless important enough to warrant an update, ensuring that we are aligning with whatever circumstances have made this course of action seem appropriate at this particular moment.
So we are changing this because of reasons that make us change it, and since those reasons exist, we have to change it. We are doing this change because things happened that made us do the change, and because of that, we are now making the change. This change is happening because the change needs to happen, and since the change is needed, we are making the change now.